backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "openai>=1.70.0",
    "pillow>=11.1.0",
    "pytest>=8.3.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "selenium>=4.31.0",
    "sqlmodel>=0.0.24",
    "webdriver-manager>=4.0.2",
]
